made seperate file for every class and corrected the private to public
