+++ Ruby-Project/capitalize.rb
@@ -1,0 +1,8 @@
+require_relative 'decorator'
+
+class CapitalizeDecorator < Decorator
+    def correct_name
+      super.capitalize
+    end
+  end
+
