+++ Ruby-Project/decorator.rb
@@ -1,0 +1,12 @@
+require_relative 'nameable_class'
+
+class Decorator < Nameable
+    def initialize(nameable)
+      super()
+      @nameable = nameable
+    end
+  
+    def correct_name
+      @nameable.correct_name
+    end
+  end
