+++ Ruby-Project/person_class.rb
@@ -15,6 +15,10 @@ class Person < Nameable
     true if @age >= 18 || @parent_permission == true
   end
 
+  def correct_name
+    @name
+  end
+
   private
 
   def of_age?
@@ -25,31 +29,7 @@ class Person < Nameable
 
     end
   end
-
-  def correct_name
-    @name
-  end
 end
 
-class Decorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
 
-  def correct_name
-    @nameable.correct_name
-  end
-end
 
-class CapitalizeDecorator < Decorator
-  def correct_name
-    super.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    super[0..9]
-  end
-end
