+++ Ruby-Project/trim.rb
@@ -1,0 +1,8 @@
+require_relative 'decorator'
+
+class TrimmerDecorator < Decorator
+    def correct_name
+      super[0..9]
+    end
+  end
+
